@@ -8,47 +8,49 @@
     "build": "vite build --mode prod",
     "build-stage": "vite build --mode stage",
     "preview": "vite preview --mode prod",
-    "preview-stage": "vite preview --mode stage",
-    "lint": "eslint . --ext js,jsx --report-unused-disable-directives --max-warnings 0"
+    "preview-stage": "vite preview --mode stage"
   },
   "dependencies": {
-    "@hookform/resolvers": "3.9.0",
-    "@radix-ui/react-dialog": "1.1.1",
-    "@radix-ui/react-label": "2.1.0",
-    "@radix-ui/react-navigation-menu": "1.2.0",
-    "@radix-ui/react-separator": "1.1.0",
-    "@radix-ui/react-slot": "1.1.0",
-    "@radix-ui/react-tabs": "1.1.1",
-    "@radix-ui/react-toast": "1.2.2",
-    "@tanstack/react-table": "8.20.5",
-    "class-variance-authority": "0.7.0",
+    "@hookform/resolvers": "3.9.1",
+    "@radix-ui/react-dialog": "1.1.4",
+    "@radix-ui/react-label": "2.1.1",
+    "@radix-ui/react-navigation-menu": "1.2.3",
+    "@radix-ui/react-separator": "1.1.1",
+    "@radix-ui/react-slot": "1.1.1",
+    "@radix-ui/react-toast": "1.2.4",
+    "@tanstack/react-query": "5.62.11",
+    "class-variance-authority": "0.7.1",
     "clsx": "2.1.1",
-    "dotenv": "16.4.5",
-    "lucide-react": "0.446.0",
+    "dotenv": "16.4.7",
+    "lucide-react": "0.469.0",
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "react-helmet-async": "2.0.5",
-    "react-hook-form": "7.53.0",
-    "react-icons": "5.3.0",
-    "react-router-dom": "6.26.2",
-    "sweetalert2": "11.14.1",
-    "tailwind-merge": "2.5.2",
-    "zod": "3.23.8"
+    "react-hook-form": "7.54.2",
+    "react-icons": "5.4.0",
+    "react-router-dom": "7.1.1",
+    "sweetalert2": "11.15.3",
+    "tailwind-merge": "2.6.0",
+    "tailwindcss-animate": "1.0.7",
+    "zod": "3.24.1",
+    "zustand": "5.0.2"
   },
   "devDependencies": {
-    "@types/react": "18.3.10",
-    "@types/react-dom": "18.3.0",
-    "@vitejs/plugin-react": "4.3.2",
+    "@eslint/js": "9.17.0",
+    "@types/node": "22.10.2",
+    "@types/react": "18.3.18",
+    "@types/react-dom": "18.3.5",
+    "@vitejs/plugin-react": "4.3.4",
     "autoprefixer": "10.4.20",
-    "eslint": "9.11.1",
-    "eslint-plugin-react": "7.37.0",
-    "eslint-plugin-react-hooks": "4.6.2",
-    "eslint-plugin-react-refresh": "0.4.12",
-    "postcss": "8.4.47",
-    "prettier": "3.3.3",
-    "prettier-plugin-tailwindcss": "0.6.8",
-    "tailwindcss": "3.4.13",
-    "tailwindcss-animate": "1.0.7",
-    "vite": "5.4.8"
+    "eslint": "9.17.0",
+    "eslint-plugin-react": "7.37.2",
+    "eslint-plugin-react-hooks": "5.0.0",
+    "eslint-plugin-react-refresh": "0.4.16",
+    "globals": "15.14.0",
+    "postcss": "8.4.49",
+    "prettier": "3.4.2",
+    "prettier-plugin-tailwindcss": "0.6.9",
+    "tailwindcss": "3.4.17",
+    "vite": "6.0.5"
   }
 }
