@@ -1,15 +1,31 @@
 from fastapi import HTTPException, status
 
 from core.logging import logger
-from core.exceptions import DatabaseError, ImageProcessError, MissingSettings
+from core.exceptions import (
+    DatabaseError,
+    ImageProcessError,
+    MissingSettings,
+    InvalidInputError,
+    NoResultsFound,
+    CacheError,
+)
 
 
 def handle_exceptions(e: Exception) -> None:
     """Handles exceptions by raising HTTPExceptions."""
 
-    if isinstance(e, (FileNotFoundError, MissingSettings, ValueError)):
+    if isinstance(
+        e,
+        (
+            FileNotFoundError,
+            MissingSettings,
+            ValueError,
+            InvalidInputError,
+            NoResultsFound,
+        ),
+    ):
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=e.args[0])
-    elif isinstance(e, (DatabaseError, ImageProcessError)):
+    elif isinstance(e, (DatabaseError, ImageProcessError, CacheError)):
         raise HTTPException(
             status_code=status.HTTP_422_UNPROCESSABLE_ENTITY, detail=e.args[0]
         )
