@@ -1,9 +1,7 @@
 from enum import Enum
 from fastapi import APIRouter
 
-from apis.v2.routers import colors
-from apis.v2.routers import files
-from apis.v2.routers import retrieve
+from apis.v2.routers import files, retrieve
 
 
 class APITag(str, Enum):
@@ -16,6 +14,6 @@ class APITag(str, Enum):
 
 router = APIRouter()
 
-router.include_router(colors.router, tags=[APITag.COLORS], prefix="/colors")
+# router.include_router(colors.router, tags=[APITag.COLORS], prefix="/colors")
 router.include_router(files.router, tags=[APITag.FILES], prefix="/upload")
 router.include_router(retrieve.router, tags=[APITag.RETRIEVE])
