@@ -1,3 +1,4 @@
+import { createElement } from "react";
 import { Link } from "react-router-dom";
 import { Outlet } from "react-router-dom";
 import { IconContext } from "react-icons";
@@ -16,19 +17,19 @@ const Home = () => {
     <IconContext.Provider value={{ size: "6em" }}>
       <div className="flex-between h-screen gap-10 bg-amber-200/50 px-12 py-7">
         {navigation_info
-          .filter((info) => info.name.toLowerCase() !== "home")
-          .map((info) => (
-            <Link key={info.name} to={info.url} className="hw-full">
+          .filter((nav) => nav.name.toLowerCase() !== "home")
+          .map((nav) => (
+            <Link key={nav.name} to={nav.url} className="hw-full">
               <Card className="flex-center hw-full hover: bg-white hover:scale-105 hover:bg-slate-50">
                 <CardHeader className="flex-center hw-full gap-3 break-words tracking-wide">
                   <CardContent className="flex-center hw-full py-16">
-                    <info.icon />
+                    {createElement(nav.icon)}
                   </CardContent>
                   <CardTitle className="hw-full text-center text-4xl font-bold">
-                    {info.title}
+                    {nav.title}
                   </CardTitle>
                   <CardDescription className="hw-full p-5 text-justify text-lg">
-                    {info.description}
+                    {nav.description}
                   </CardDescription>
                 </CardHeader>
               </Card>
