@@ -1,25 +1,31 @@
 import { defineConfig, loadEnv } from "vite";
-import path from "path";
-import dotenv from "dotenv";
+import { resolve, dirname } from "path";
+import { fileURLToPath } from "url";
 import react from "@vitejs/plugin-react";
+import process from "process";
+import dotenv from "dotenv";
 
-dotenv.config({ path: "../.env" }); // load env vars from .env
+dotenv.config({ path: "../.env" });
+
+// Utility for ESM compatibility
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
 
 // https://vitejs.dev/config/
-export default defineConfig(({ _, mode }) => {
+export default defineConfig(({ mode }) => {
   // Load environment variables based on the current mode.
-  const env = { ...process.env, ...loadEnv(mode, process.cwd()) };
+  const env = loadEnv(mode, process.cwd());
 
   return {
     plugins: [react()],
     resolve: {
       alias: {
-        "@": path.resolve(__dirname, "./src"),
+        "@": resolve(__dirname, "./src"),
       },
     },
     server: {
       host: true,
-      port: env.VITE_APP_PORT,
+      port: parseInt(env.VITE_APP_PORT) || 3000,
       strictPort: true,
       proxy: {
         "/api": {
@@ -30,7 +36,7 @@ export default defineConfig(({ _, mode }) => {
       },
     },
     preview: {
-      port: env.VITE_APP_PORT,
+      port: parseInt(env.VITE_APP_PORT) || 5000,
     },
   };
 });
